Add Dashboard and AdminComponent components
