+++ hosting/src/components/AdminComponent.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const AdminComponent = () => (
+    <div className = "adminComponent">
+      { "sorry nothing" }
+    </div>
+  );
+  
+  export default AdminComponent;
